requirements.txt
flask==2.0.2
azure-cosmos==4.3.0